node:https: forward minVersion/maxVersion/secureProtocol to the server TLS context
https.createServer() ran its TLS options through a hardcoded subset when
building the Bun.serve({ tls }) config in _http_server.ts, so the protocol
version bounds never reached the native context. A minVersion:'TLSv1.3'
server would happily negotiate TLS1.2, and a maxVersion:'TLSv1.2' server
would negotiate TLS1.3, in both directions, with no error.

Move the minVersion/maxVersion/secureProtocol resolution and validation
into internal/tls.ts so node:tls and the node:http(s) server share one
implementation, and forward the resolved integer range from the https
Server constructor.
